Replaced - (unsigned int) count by - (NSUInteger) count in Inbox/Pantomime to prevent return type conflict with GNUstep Foundation classes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6129 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/User/Inbox/Pantomime/Framework/Pantomime/CWCacheManager.m
@@ -82,7 +82,7 @@
 //
 //
 //
-- (unsigned int) count
+- (NSUInteger) count
 {
   return _count;
 }
