+++ Services/User/Inbox/Pantomime/Framework/Pantomime/CWContainer.m
@@ -170,7 +170,7 @@
 //
 //
 //
-- (unsigned int) count
+- (NSUInteger) count
 {
   if (child)
     {
